architecture.md: Write serialization layer section
The scope this section covers is not yet determined. For the moment it
only check for syntax errors in the instructions themselves, but we
could extend this concept to a conversion from instructions to
subroutines in the implementation language. We will see where it leads
us in the next 2 sections, maybe that the extent of the instructions
serialization will broaden.