Improve tests for instruction FX0A.
In the previous tests, the behaviour of the instruction when no key was
pressed was not yet defined. Now, in a such case, the FX0A instruction
should decrement the engine's program counter in order to make the next
engine cycle re-read the FX0A instruction, until a key is pressed.