chip-8.md: Fix description of instruction 2NNN
The previous explanation was saying that before the value of the program
counter was stored inside the stack, it should be incremented by 2 in
order to point to the next instruction when returning from the
subroutine.
This is wrong. The engine cycle automatically increments the program
counter by itself, so this is unneeded.