Fix engine init tests and program loading tests
Change the default jest regex to include all *.js files inside test/
directory.
In the 00-engine-initialization.js tests, some value for checking that
an engine component is X bytes long were wrong. This is fixed.
In the 03-program-loading.js tests, there were an error of memory
parsing when comparing the program to load and it representation inside
the memory. Also, in src/engine.js, a wrong index inside the
loadProgram() was used to go through the program. This is now fixed and
program loading tests now pass.