utility.md: Fix design documentation to match code
Writing the utility module led to some adjustements regarding this
documentation concerning the responsability of certain functions.

An example is the dump/load tools. The doc said that they managed the
files on the user's FS concerning the dumps. A such thing shouldn't be
their responsability, it is the one of the UIs.

And removed the purifyFile() function from the doc.

Also changed the compareEngines tool to display all values in decimal
(indexes weren't).