cli.js: Fix API use of the utility module
The CLI was using the routines provided by the utility, mainly the
load/dump tools, with delegating the FS operations to the utility
module. But due to recent changes to the utility module, it is now the
role of the CLI to manage the files writings/readings.

Changes were made to match this new behaviour.